chore: replace `useModalModelEdit` with `ModelEditModal`
