+++ webui/react/src/components/ModelEditModal.tsx
@@ -1,0 +1,66 @@
+import Form from 'components/kit/Form';
+import Input from 'components/kit/Input';
+import { Modal } from 'components/kit/Modal';
+import { patchModel } from 'services/api';
+import { ModelItem } from 'types';
+import handleError from 'utils/error';
+
+type FormInputs = {
+  modelName: string;
+  description?: string;
+};
+
+interface Props {
+  fetchModel: () => Promise<void>;
+  model: ModelItem;
+}
+
+const FORM_ID = 'edit-model-form';
+
+const ModelEditModal = ({ model, fetchModel }: Props): JSX.Element => {
+  const [form] = Form.useForm<FormInputs>();
+
+  const handleOk = async () => {
+    const values = await form.validateFields();
+    try {
+      await patchModel({
+        body: { description: values.description, name: values.modelName },
+        modelName: model.name,
+      });
+      await fetchModel();
+    } catch (e) {
+      handleError(e, {
+        publicSubject: 'Unable to edit model',
+        silent: false,
+      });
+    }
+  };
+
+  const handleClose = () => {
+    form.resetFields();
+  };
+
+  return (
+    <Modal
+      okButtonProps={{ form: FORM_ID, htmlType: 'submit' }}
+      size="small"
+      submit={{ handler: handleOk, text: 'Save' }}
+      title="Edit Model"
+      onClose={handleClose}>
+      <Form autoComplete="off" form={form} id={FORM_ID} layout="vertical">
+        <Form.Item
+          initialValue={model.name}
+          label="Name"
+          name="modelName"
+          rules={[{ message: 'Name is required', required: true }]}>
+          <Input />
+        </Form.Item>
+        <Form.Item initialValue={model.description} label="Description" name="description">
+          <Input />
+        </Form.Item>
+      </Form>
+    </Modal>
+  );
+};
+
+export default ModelEditModal;
